init offer and price models

+++ src/Traits/ClinicEditActions.php
@@ -120,6 +120,12 @@ trait ClinicEditActions
         if (! $modelObject) { return; }
         if (! $this->checkAuth("delete", $modelObject)) { return; }
 
+        if ($modelObject->offers()->count() > 0) {
+            session()->flash("error", "Невозможно удалить клинику, есть предложения");
+            $this->closeDelete();
+            return;
+        }
+
         try {
             $modelObject->delete();
         } catch (\Exception $exception) {
